Add auto pull, refactor extension.ts
