+++ CHANGELOG.md
@@ -4,11 +4,15 @@ All notable changes to the "multi-repo-checkout" extension will be documented in
 
 ## Released
 
+### [0.1.8]
+- Enhancement [Auto Pull after checkout if possible](https://github.com/wolframs/multi-repo-checkout/issues/10)
+- Move functional code out of extension.ts to separate .ts files
+
 ### [0.1.7]
 - Enhancement [Add screenshots / gif of functionality to README.md](https://github.com/wolframs/multi-repo-checkout/issues/3)
 
 ### [0.1.6]
-- Feature [Auto reload window after succesful checkouts (with settings flag)](https://github.com/wolframs/multi-repo-checkout/issues/2)
+- Feature [Auto reload window after successful checkouts (with settings flag)](https://github.com/wolframs/multi-repo-checkout/issues/2)
 - Added configuration to configure "Registering changes..." delay
 
 ### [0.1.5]
